ducdv_fix

+++ app/Services/Typecar/TypeCarService.php
@@ -4,14 +4,10 @@ namespace App\Services\Typecar;
 use App\Http\Requests\Car\UpdateCarRequest;
 use App\Http\Requests\TypeCar\StoreTypeCarRequest;
 use App\Models\TypeCar;
+use App\Models\Car;
 
 class TypeCarService
     {
-//    public function index(){
-//        $title = 'Trang phân quyền';
-//        $data = TypeCar::query()->get();
-//        return view('admin.pages.typecar.main',compact('title','data'));
-//    }
     public function store($request){
         $model = new TypeCar();
         $model->fill($request->all());
@@ -25,6 +21,14 @@ class TypeCarService
     public function destroy(string $id)
     {
         $model = TypeCar::query()->findOrFail($id);
+        $relatedCars = Car::where('id_type_car', $id)->count();
+
+        if ($relatedCars > 0) {
+            // Nếu có bản ghi liên quan, bạn không thể xóa TypeCar
+            log(123);
+            return redirect()->back();
+        }
+
         $model->delete();
 
     }
